+++ src/IRProtocol.hpp
@@ -74,6 +74,7 @@ const char string_Lego[] PROGMEM = "Lego";
 const char string_MagiQuest[] PROGMEM = "MagiQuest";
 const char string_Whynter[] PROGMEM = "Whynter";
 const char string_FAST[] PROGMEM = "FAST";
+const char string_Other[] PROGMEM = "OTHER";
 
 /*
  * !!Must be the same order as in decode_type_t in IRProtocol.h!!!
@@ -84,7 +85,7 @@ PROGMEM = { string_Unknown, string_PulseWidth, string_PulseDistance, string_Appl
         string_Kaseikyo_JVC, string_Kaseikyo_Mitsubishi, string_RC5, string_RC6, string_RC6A, string_Samsung, string_SamsungLG,
         string_Samsung48, string_Sharp, string_Sony
 #if !defined(EXCLUDE_EXOTIC_PROTOCOLS)
-        , string_BangOlufsen, string_BoseWave, string_Lego, string_MagiQuest, string_Whynter, string_FAST
+        , string_BangOlufsen, string_BoseWave, string_Lego, string_MagiQuest, string_Whynter, string_FAST, string_Other
 #endif
         };